requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9

google-api-python-client==2.141.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.4

docxtpl==0.16.7
python-docx==1.1.2
Pillow==10.4.0

python-dotenv==1.0.1

matplotlib==3.9.2

docx2pdf==0.1.8
